0-apps/xml-reader-producer/requirements.txt
# general libs
xmltodict==0.12.0
confluent-kafka==1.6.0
python-dotenv==0.15.0
azure-storage-blob==12.8.0
python-dateutil==2.8.1
psycopg2==2.8.6
CMRESHandler==1.0.0
elasticsearch==7.11.0
requests==2.25.1

# documentation
Sphinx==3.5.2
Flask-Sphinx-Themes==1.0.2
rinohtype==0.5.2

# linting and identation
black==20.8b1
pylint==2.7.2